python: change test scripts to work on Python 2.4

+++ python/test/helpers_test.py
@@ -3,7 +3,6 @@
 ### $Copyright: copyright(c) 2010-2011 kuwata-lab.com all rights reserved $
 ### $License: MIT License $
 ###
-from __future__ import with_statement
 
 import sys, os, re
 import unittest
@@ -13,6 +12,21 @@ from oktest.helper import *
 from oktest.helper import flatten, rm_rf
 from oktest.dummy import *
 
+available_with_statement = sys.version_info[0:2] > (2, 4)
+
+
+def do_exec(script, **kwargs):
+    if not available_with_statement:
+        print("*** skipped")
+        return
+    script = "from __future__ import with_statement\n" + script + "\n"
+    gvars = {'sys': sys, 'os': os, 're': re, 'ok': ok, 'NG': NG}
+    for k in oktest.dummy.__all__:
+        gvars[k] = getattr(oktest.dummy, k)
+    gvars.update(kwargs)
+    exec(script, gvars, gvars)
+
+
 
 class _Context_TC(unittest.TestCase):
 
@@ -20,11 +34,11 @@ class _Context_TC(unittest.TestCase):
         self.ctx = oktest._Context()
 
     def test___enter__(self):
-        with spec("returns self."):
+        if spec("returns self."):
             ok (self.ctx.__enter__()).is_(self.ctx)
 
     def test__exit__(self):
-        with spec("returns None."):
+        if spec("returns None."):
             ok (self.ctx.__exit__()) == None
 
 
@@ -34,7 +48,7 @@ class _RunnableContext_TC(unittest.TestCase):
         ctx = oktest._RunnableContext()
         ret = None
         #
-        with spec("calls __enter__() and __exit__() to emurate with-statement."):
+        if spec("calls __enter__() and __exit__() to emurate with-statement."):
             called = []
             def enter(*args):
                 called.append(('enter', args))
@@ -52,18 +66,18 @@ class _RunnableContext_TC(unittest.TestCase):
             ok (called[1]) == ('func', ('a',), {'b': 1})
             ok (called[2]) == ('exit', (None, None, None))
         #
-        with spec("returns value which func returned."):
+        if spec("returns value which func returned."):
             ok (ret) == 123
 
     def test_deco(self):
         ctx = oktest._RunnableContext()
-        with spec("decorates function."):
+        if spec("decorates function."):
             @ctx.deco
             def func(x):
                 return x + 10
             ret = func(123)
             ok (ret) == 133
-        with spec("__enter__() and __exit__() are called when decoreated function called."):
+        if spec("__enter__() and __exit__() are called when decoreated function called."):
             called = []
             def enter(*args): called.append(('enter', args))
             def exit(*args): called.append(('exit', args))
@@ -84,12 +98,12 @@ class _RunnableContext_TC(unittest.TestCase):
 class Spec_TC(unittest.TestCase):
 
     def test___init__(self):
-        with spec("takes description."):
+        if spec("takes description."):
             obj = oktest.Spec('foobar')
             ok (obj.desc) == 'foobar'
 
     def test___iter__(sefl):
-        with spec("emurates with-stmt when used with for-stmt."):
+        if spec("emurates with-stmt when used with for-stmt."):
             called = []
             def enter(*args):
                 called.append(('enter', args))
@@ -111,10 +125,10 @@ class Spec_TC(unittest.TestCase):
     def test___bool__(self):
         obj = oktest.Spec('SOS')
         #
-        with spec("returns True when $SPEC is not set."):
+        if spec("returns True when $SPEC is not set."):
             ok (bool(obj)) == True
         #
-        with spec("returns False when $SPEC not matched to SPEC."):
+        if spec("returns False when $SPEC not matched to SPEC."):
             try:
                 os.environ['SPEC'] = 'SOS'
                 ok (bool(obj)) == True
@@ -130,9 +144,9 @@ class GLOBAL_TC(unittest.TestCase):
 
     def test_spec(self):
         obj = oktest.spec('Haruhi')
-        with spec("returns oktest.Spec object."):
+        if spec("returns oktest.Spec object."):
             ok (obj).is_a(oktest.Spec)
-        with spec("takes description."):
+        if spec("takes description."):
             ok (obj.desc) == 'Haruhi'
 
 
@@ -149,14 +163,17 @@ class helper_TC(unittest.TestCase):
             cwd = os.getcwd()
             tmpdir_path = os.path.abspath(tmpdir)
             #+
-            with spec("change directory temporarily when used with with-stmt."):
+            if spec("change directory temporarily when used with with-stmt."):
+                script = r"""if True:
+                import oktest
                 with oktest.helper.chdir(tmpdir):
                     ok (os.getcwd()) != cwd
                     ok (os.getcwd()) == tmpdir_path
-                ok (os.getcwd()) == cwd
+                ok (os.getcwd()) == cwd  """
+                do_exec(script, tmpdir=tmpdir, cwd=cwd, tmpdir_path=tmpdir_path)
             #-
             #
-            with spec("change directory temporarily when run() called with function."):
+            if spec("change directory temporarily when run() called with function."):
                 def f():
                     ok (os.getcwd()) != cwd
                     ok (os.getcwd()) == tmpdir_path
@@ -169,19 +186,23 @@ class helper_TC(unittest.TestCase):
 
     def test_using(self):
         #+
-        with spec("takes a class object and adds local vars to it when used with with-stmt."):
+        if spec("takes a class object and adds local vars to it when used with with-stmt."):
+            script = r"""if True:
+            import oktest
             with oktest.helper.using(DummyClass) as obj:
                 def test_1(self):
                     ok (1+1) == 2
                 def test_2(self):
-                    ok (1-1) == 0
-            ok (DummyClass).has_attr('test_1')
-            ok (DummyClass).has_attr('test_2')
+                    ok (1-1) == 0  """
+            do_exec(script, DummyClass=DummyClass)
+            if available_with_statement:
+                ok (DummyClass).has_attr('test_1')
+                ok (DummyClass).has_attr('test_2')
         #-
         pass
 
     def test_flatten(self):
-        with spec("flatten nested list or tuple."):
+        if spec("flatten nested list or tuple."):
             ret = oktest.helper.flatten([[1, (2, (3, 4)), 5]])
             ok (ret) == [1, 2, 3, 4, 5]
             ret = flatten([1, [2, 3, [4, 5, [[[6]]]], [7, 8]]])
@@ -198,13 +219,15 @@ class helper_TC(unittest.TestCase):
                 f.close()
                 ok (fname).is_file()
             #
-            with spec("removes files and directories recursively."):
+            if spec("removes files and directories recursively."):
                 oktest.helper.rm_rf('_test_foo1', '_test_foo2', '_test_bar9')
                 ok ('_test_foo1').not_exist()
                 ok ('_test_foo2').not_exist()
                 ok ('_test_bar9').not_exist()
         finally:
-            pass
+            for x in ['_test_foo1', '_test_foo2']:
+                if os.path.isdir(x):
+                    shutil.rmtree(x)
 
 
 class helper_rm_rf_TC(unittest.TestCase):
@@ -242,20 +265,27 @@ class helper_rm_rf_TC(unittest.TestCase):
 
 from oktest.dummy import *
 
+def new_gvars(**kwargs):
+    d = {'sys': sys, 'os': os, 're': re, 'ok': ok, 'NG': NG}
+    d.update(kwargs)
+    return d
+
+
 class dummy_TC(unittest.TestCase):
+
 
     def test_dummy_file(self):
         fname, content = '_test.sos.txt', 'SOS'
-        #+
-        with spec("creates a dummy file temporarily when used with with-stmt."):
+        if spec("creates a dummy file temporarily when used with with-stmt."):
+            script = r"""if True:
             with dummy_file(fname, content):
                 ok (fname).is_file()
                 f = open(fname); s = f.read(); f.close()
                 ok (s) == content
-            NG (fname).is_file()
-        #-
+            NG (fname).is_file()  """
+            do_exec(script, fname=fname, content=content)
         #
-        with spec("creates a dummy file temporarily when run() called with a function."):
+        if spec("creates a dummy file temporarily when run() called with a function."):
             called = []
             def func():
                 called.append(('func', fname))
@@ -268,7 +298,7 @@ class dummy_TC(unittest.TestCase):
             ok (called[0]) == ('func', fname)
             ok (ret) == 999
         #
-        with spec("available as decorator."):
+        if spec("available as decorator."):
             called = []
             @dummy_file(fname, content)
             def func():
@@ -285,13 +315,15 @@ class dummy_TC(unittest.TestCase):
     def test_dummy_dir(self):
         dname = '_test.sos.dir'
         #+
-        with spec("creates a dummy directory temporarily when used with with-stmt."):
+        if spec("creates a dummy directory temporarily when used with with-stmt."):
+            script = r"""if True:
             with dummy_dir(dname):
                 ok (dname).is_dir()
-            NG (dname).is_dir()
+            NG (dname).is_dir()  """
+            do_exec(script, dname=dname)
         #-
         #
-        with spec("creates a dummy directory temporarily when run() called with a function."):
+        if spec("creates a dummy directory temporarily when run() called with a function."):
             called = []
             def func():
                 called.append(('func', dname))
@@ -302,7 +334,7 @@ class dummy_TC(unittest.TestCase):
             ok (called[0]) == ('func', dname)
             ok (ret) == 888
         #
-        with spec("available as decorator."):
+        if spec("available as decorator."):
             called = []
             @dummy_dir(dname)
             def func():
@@ -317,14 +349,16 @@ class dummy_TC(unittest.TestCase):
     def test_dummy_values(self):
         d = {'Haruhi': 'Suzumiya'}
         #+
-        with spec("changes dictionary value temporarily when used with with-stmt."):
+        if spec("changes dictionary value temporarily when used with with-stmt."):
+            script = r"""if True:
             with dummy_values(d, Yuki='Nagato'):
                 ok (d.get('Yuki', None)) == 'Nagato'
             ok ('Haruhi' in d) == True
-            ok ('Yuki' in d) == False
+            ok ('Yuki' in d) == False  """
+            do_exec(script, d=d)
         #-
         #
-        with spec("changes dictionary value temporarily when run() called with function."):
+        if spec("changes dictionary value temporarily when run() called with function."):
             called = []
             def func():
                 called.append(True)
@@ -333,7 +367,7 @@ class dummy_TC(unittest.TestCase):
             ok (called) == [True]
             ok (d) == {'Haruhi': 'Suzumiya'}
         #
-        with spec("available as decorator."):
+        if spec("available as decorator."):
             called = []
             @dummy_values(d, Mikuru='Asahina')
             def func():
@@ -345,14 +379,16 @@ class dummy_TC(unittest.TestCase):
     def test_dummy_attrs(self):
         obj = DummyClass()
         #+
-        with spec("changes attributes temporarily when used with with-stmt."):
+        if spec("changes attributes temporarily when used with with-stmt."):
+            script = r"""if True:
             with dummy_attrs(obj, SOS=[123]):
                 ok (obj).has_attr('SOS')
                 ok (obj.SOS) == [123]
-            NG (obj).has_attr('SOS')
+            NG (obj).has_attr('SOS')  """
+            do_exec(script, obj=obj)
         #-
         #
-        with spec("changes attributes temporarily when run() called with function."):
+        if spec("changes attributes temporarily when run() called with function."):
             called = []
             def func():
                 called.append(True)
@@ -364,7 +400,7 @@ class dummy_TC(unittest.TestCase):
             NG (obj).has_attr('SOS')
             ok (ret) == 999
         #
-        with spec("available as decorator."):
+        if spec("available as decorator."):
             called = []
             @dummy_attrs(obj, SOS=[123])
             def func():
@@ -379,15 +415,17 @@ class dummy_TC(unittest.TestCase):
 
     def test_dummy_environ_vars(self):
         #+
-        with spec("changes environment variables temporarily when used with with-stmt."):
+        if spec("changes environment variables temporarily when used with with-stmt."):
+            script = r"""if True:
             ok ('SOS' in os.environ) == False
             with dummy_environ_vars(SOS='!!!'):
                 ok ('SOS' in os.environ) == True
                 ok (os.environ['SOS']) == '!!!'
-            ok ('SOS' in os.environ) == False
+            ok ('SOS' in os.environ) == False  """
+            do_exec(script)
         #-
         #
-        with spec("changes environment variables temporarily when run() called with function."):
+        if spec("changes environment variables temporarily when run() called with function."):
             called = []
             def func():
                 called.append(True)
@@ -399,7 +437,7 @@ class dummy_TC(unittest.TestCase):
             ok ('SOS' in os.environ) == False
             ok (ret) == 111
         #
-        with spec("available as decorator."):
+        if spec("available as decorator."):
             called = []
             @dummy_environ_vars(SOS='???')
             def func():
@@ -417,7 +455,8 @@ class dummy_TC(unittest.TestCase):
         sout = sys.stdout
         serr = sys.stderr
         #+
-        with spec("changes stdio temporarily when used with with-stmt."):
+        if spec("changes stdio temporarily when used with with-stmt."):
+            script = r"""if True:
             with dummy_io("foobar") as d_io:
                 NG (sys.stdin).is_(sin)
                 NG (sys.stdout).is_(sout)
@@ -430,10 +469,11 @@ class dummy_TC(unittest.TestCase):
             ok (sys.stdout).is_(sout)
             ok (sys.stderr).is_(serr)
             ok (d_io.stdout) == "Haruhi\nSasaki\n"
-            ok (d_io.stderr) == "Kyon"
+            ok (d_io.stderr) == "Kyon"  """
+            do_exec(script, sin=sin, sout=sout, serr=serr)
         #-
         #
-        with spec("changes stdio temporarily when run() called with function."):
+        if spec("changes stdio temporarily when run() called with function."):
             called = []
             def func():
                 called.append(True)
@@ -455,7 +495,7 @@ class dummy_TC(unittest.TestCase):
             ok (d_io.stderr) == "Itsuki"
             ok (ret) == 777
         #
-        with spec("available as decorator."):
+        if spec("available as decorator."):
             called = []
             @dummy_io("SOS")
             def func():
